+++ python/GafferSceneUI/LightEditor.py
@@ -314,42 +314,6 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 				}
 			)
 
-			menuDefinition.append( "/deleteDivider", { "divider" : True } )
-
-			# Filter out a number of scenarios where deleting would be impossible
-			# or unintuitive
-			deleteEnabled = True
-			inputNode = self.settings()["in"].getInput().node()
-			editScopeNode = self.editScope()
-			if editScopeNode is not None :
-				if inputNode != editScopeNode and editScopeNode not in Gaffer.NodeAlgo.upstreamNodes( inputNode ) :
-					# Edit scope is downstream of input
-					deleteEnabled = False
-				elif GafferScene.EditScopeAlgo.prunedReadOnlyReason( editScopeNode ) is not None :
-					# Pruning or the edit scope is read only
-					deleteEnabled = False
-				else :
-					with self.context() :
-						if not editScopeNode["enabled"].getValue() :
-							# Edit scope is disabled
-							deleteEnabled = False
-						else :
-							pruningProcessor = editScopeNode.acquireProcessor( "PruningEdits", createIfNecessary = False )
-							if pruningProcessor is not None and not pruningProcessor["enabled"].getValue() :
-								# Pruning processor is disabled
-								deleteEnabled = False
-			else :
-				# No edit scope selected
-				deleteEnabled = False
-
-			menuDefinition.append(
-				"Delete",
-				{
-					"command" : Gaffer.WeakMethod( self.__deleteLights ),
-					"active" : deleteEnabled
-				}
-			)
-
 	def __selectLinked (self, *unused ) :
 
 		context = self.context()
@@ -370,16 +334,6 @@ class LightEditor( GafferSceneUI.SceneEditor ) :
 
 		if not isinstance( result, Exception ) :
 			GafferSceneUI.ScriptNodeAlgo.setSelectedPaths( self.scriptNode(), result )
-
-	def __deleteLights( self, *unused ) :
-
-		# There may be multiple columns with a selection, but we only operate on the name column.
-		selection = self.__pathListing.getSelection()[0]
-
-		editScope = self.editScope()
-
-		with Gaffer.UndoScope( editScope.ancestor( Gaffer.ScriptNode ) ) :
-			GafferScene.EditScopeAlgo.setPruned( editScope, selection, True )
 
 GafferUI.Editor.registerType( "LightEditor", LightEditor )