+++ Changes.md
@@ -30,6 +30,7 @@ Improvements
   - Allow outputs with layer names and outputs without layer names to be merged to the same file.
 - Viewer : Added `Hide` and `Prune` menu items to the right-click menu of the 3D view.
 - LightEditor, AttributeEditor : Added `Hide` and `Prune` menu items.
+- LightEditor : Removed `Delete` menu item, use `Prune` instead.
 
 Fixes
 -----